Move config into its own section
There's probably a few ways to do this config handling better but for
now just isolate it into its own file. Includes the loggers which is
probably not the right way to do this but whatever.
