+++ index.js
@@ -3,63 +3,10 @@
 const CIDRMatcher = require('cidr-matcher');
 const http = require('http');
 const net = require('net');
-const node_static = require('node-static');
-const path = require('path');
 const querystring = require('querystring');
 const sockjs = require('sockjs');
-const winston = require('winston');
-const winstonConf = require('winston-config');
-const yargs = require('yargs');
 
-var argv = yargs
-    .usage('Usage: $0 <command> [options]')
-    .env('BLP')
-    .option('l', {
-        alias: 'log-config',
-        nargs: 1,
-        describe: 'Load winston config logging from file',
-        normalize: true,
-        coerce: path.resolve
-    })
-    .option('p', {
-        alias: 'port',
-        nargs: 1,
-        describe: 'The port to listen on',
-        number: true,
-        default: 3000
-    })
-    .option('b', {
-        alias: 'bind',
-        nargs: 1,
-        describe: 'The interface IP to bind to',
-        string: true,
-        default: '0.0.0.0'
-    })
-    .help('h')
-    .alias('h', 'help')
-    .argv;
-
-if (argv.l) {
-    var logger = winstonConf.fromFileSync(argv.l);
-
-    var appLogger = logger.loggers.get('application');
-    var telnetLogger = logger.loggers.get('telnet');
-} else {
-    var logger = new winston.Logger({
-        transports: [
-            new winston.transports.Console({
-                level: 'debug',
-                handleExceptions: true,
-                json: false,
-                colorize: true
-            })
-        ],
-        exitOnError: true
-    });
-
-    var appLogger = logger;
-    var telnetLogger = logger;
-}
+const config = require('./config');
 
 var matcher = new CIDRMatcher([ '10.0.0.0/8', '172.16.30.0/24' ]);
 
@@ -72,7 +19,7 @@ sockjs_server.on('connection', function(conn) {
     loc = conn.url.indexOf("?");
 
     if (loc < 0) {
-        appLogger.info("no query string found");
+        config.logger.app.info("no query string found");
         conn.close()
         return;
     }
@@ -82,13 +29,13 @@ sockjs_server.on('connection', function(conn) {
     host = params["host"];
 
     if (!host) {
-        appLogger.info('host not specified in query string: ' + qs);
+        config.logger.app.info('host not specified in query string: ' + qs);
         conn.close()
         return;
     }
 
     if (!matcher.contains(host)) {
-        appLogger.warning("Attempt to proxy to an unauthorized IP: " + host);
+        config.logger.app.warning("Attempt to proxy to an unauthorized IP: " + host);
         conn.close();
         return;
     }
@@ -97,7 +44,7 @@ sockjs_server.on('connection', function(conn) {
         // this callback gets triggered when a successful connection is established
 
         client.on('close', function(had_error) {
-            telnetLogger.debug("telnet connection closed");
+            config.logger.telnet.debug("telnet connection closed");
             conn.close();
         });
 
@@ -116,16 +63,16 @@ sockjs_server.on('connection', function(conn) {
     client.on('error', function(err) {
         // this handler is outside the connect callback to handle errors before
         // connect occurs
-        telnetLogger.warning("Error: " + err.message);
+        config.logger.telnet.warning("Error: " + err.message);
         conn.write('error: ' +  err.message);
         conn.end()
     });
 
-    appLogger.debug("setup complete to IP: " + host);
+    config.logger.app.debug("setup complete to IP: " + host);
 
     conn.on('data', function(message) {
         // we shouldn't receive any information from the websocket.
-        telnetLogger.warning("data received on websocket: " + message);
+        config.logger.telnet.warning("data received on websocket: " + message);
     });
 
     conn.on('close', function() {
@@ -144,5 +91,5 @@ server.addListener('upgrade', function(req,res) {
 
 sockjs_server.installHandlers(server, { prefix: '/sock' });
 
-appLogger.info('listening on ' + argv.b + ':' + argv.p);
-server.listen(argv.p, argv.b);
+config.logger.app.info('listening on ' + config.bind_host + ':' + config.port);
+server.listen(config.port, config.bind_host);
